Unified the API, added ICA denoising function
Arguments now have the same names across functions. Matplotlib was dropped as a dependency (turns out we didn't actually use it anymore, oops!). Scikit-learn was added as a dependency, as the ICA denoising was now added as a function. (Previously left out to avoid having to add sklearn as dependency.)